refactor: prune unused graph-first warp surface
